requirements.txt
pymongo
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask-SQLAlchemy
Flask-Login
waitress